extracted more from the base command to other places
