+++ src/Command/BaseRokkaCliCommand.php
@@ -2,22 +2,13 @@
 
 namespace RokkaCli\Command;
 
-use Rokka\Client\Core\DynamicMetadata\DynamicMetadataInterface;
-use Rokka\Client\Core\DynamicMetadata\SubjectArea;
-use Rokka\Client\Core\Membership;
-use Rokka\Client\Core\Organization;
-use Rokka\Client\Core\SourceImage;
-use Rokka\Client\Core\Stack;
 use Rokka\Client\Image;
-use RokkaCli\Configuration;
 use RokkaCli\Provider\ClientProvider;
-use RokkaCli\RokkaHelper;
-use RokkaCli\RokkaLibrary;
+use RokkaCli\RokkaApiHelper;
+use RokkaCli\RokkaFormatter;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Helper\FormatterHelper;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
-use Symfony\Component\Yaml\Yaml;
 
 abstract class BaseRokkaCliCommand extends Command
 {
@@ -27,20 +18,20 @@ abstract class BaseRokkaCliCommand extends Command
     protected $clientProvider;
 
     /**
-     * @var RokkaHelper
+     * @var RokkaApiHelper
      */
     protected $rokkaHelper;
 
     /**
-     * @var FormatterHelper
+     * @var RokkaFormatter
      */
     protected $formatterHelper = null;
 
-    public function __construct(ClientProvider $clientProvider, RokkaHelper $rokkaHelper)
+    public function __construct(ClientProvider $clientProvider, RokkaApiHelper $rokkaHelper)
     {
         $this->clientProvider = $clientProvider;
         $this->rokkaHelper = $rokkaHelper;
-        $this->formatterHelper = new FormatterHelper();
+        $this->formatterHelper = new RokkaFormatter();
         parent::__construct();
     }
 
@@ -49,13 +40,13 @@ abstract class BaseRokkaCliCommand extends Command
      */
     protected function initialize(InputInterface $input, OutputInterface $output)
     {
-        if ($this->rokkaHelper->isDefaultApiUri()) {
+        if ($this->clientProvider->isDefaultApiUri()) {
             return;
         }
 
         $output->writeln($this->formatterHelper->formatBlock([
             'Warning!',
-            'Rokka API Uri has been overridden, API calls are performed to "'.$this->rokkaHelper->getApiUri().'".',
+            'Rokka API Uri has been overridden, API calls are performed to "'.$this->clientProvider->getApiUri().'".',
         ], 'comment', true));
     }
 
@@ -180,139 +171,5 @@ abstract class BaseRokkaCliCommand extends Command
         }
 
         return true;
-    }
-
-
-    /**
-     * @param string $fileName
-     * @param Configuration $config
-     *
-     * @return int|bool
-     */
-    protected function updateConfigToFile($fileName, Configuration $config)
-    {
-        $configArray = [
-            'api_key' => $config->getApiKey(),
-            'api_secret' => $config->getApiSecret(),
-            'api_uri' => $config->getApiUri(),
-            'organization' => $config->getOrganizationName() ? $config->getOrganizationName() : '',
-        ];
-
-        $yml = [];
-        if (file_exists($fileName)) {
-            $yml = Yaml::parse(file_get_contents($fileName));
-        }
-
-        $yml = array_merge($yml, ['rokka_cli' => $configArray]);
-
-        return file_put_contents($fileName, Yaml::dump($yml));
-    }
-
-    /**
-     * @param SourceImage     $sourceImage
-     * @param OutputInterface $output
-     */
-    public static function outputImageInfo(SourceImage $sourceImage, OutputInterface $output)
-    {
-        $output->writeln([
-            '  Hash: <info>'.$sourceImage->hash.'</info>',
-            '  Organization: <info>'.$sourceImage->organization.'</info>',
-            '  Name: <info>'.$sourceImage->name.'</info>',
-            '  Size: <info>'.$sourceImage->size.'</info>',
-            '  Format: <info>'.$sourceImage->format.'</info>',
-            '  Created: <info>'.$sourceImage->created->format('Y-m-d H:i:s').'</info>',
-            '  Dimensions: <info>'.$sourceImage->width.'x'.$sourceImage->height.'</info>',
-        ]);
-
-        if ($output->isVerbose()) {
-            $output->writeln('  BinaryHash: <info>'.$sourceImage->binaryHash.'</info>');
-        }
-
-        if (!empty($sourceImage->dynamicMetadata)) {
-            if (!$output->isVerbose()) {
-                $metaNames = array_keys($sourceImage->dynamicMetadata);
-                $output->writeln('  DynamicMetadatas ('.count($metaNames).'): '.implode(', ', $metaNames));
-            } else {
-                $output->writeln('  DynamicMetadatas:');
-                /** @var DynamicMetadataInterface $meta */
-                foreach ($sourceImage->dynamicMetadata as $name => $meta) {
-                    $output->writeln('     - <info>'.$name.'</info> '.self::getDynamicMetadataInfo($meta));
-                }
-            }
-        }
-    }
-
-    /**
-     * @param Organization    $org
-     * @param OutputInterface $output
-     */
-    public static function outputOrganizationInfo(Organization $org, OutputInterface $output)
-    {
-        $output->writeln('  ID: <info>'.$org->getId().'</info>');
-        $output->writeln('  Name: <info>'.$org->getName().'</info>');
-        $output->writeln('  Display Name: <info>'.$org->getDisplayName().'</info>');
-        $output->writeln('  Billing eMail: <info>'.$org->getBillingEmail().'</info>');
-    }
-
-    /**
-     * @param Membership      $membership
-     * @param OutputInterface $output
-     */
-    public static function outputOrganizationMembershipInfo(Membership $membership, OutputInterface $output)
-    {
-        $output->writeln('  ID: <info>'.$membership->userId.'</info>');
-        $output->writeln('  Role: <info>'.$membership->role.'</info>');
-        $output->writeln('  Active: <info>'.($membership->active ? 'True' : 'False').'</info>');
-    }
-
-    /**
-     * @param Stack           $stack
-     * @param OutputInterface $output
-     */
-    public static function outputStackInfo(Stack $stack, OutputInterface $output)
-    {
-        $output->writeln('  Name: <info>'.$stack->getName().'</info>');
-        $output->writeln('  Created: <info>'.$stack->getCreated()->format('Y-m-d H:i:s').'</info>');
-
-        $operations = $stack->getStackOperations();
-        if (!empty($operations)) {
-            $output->writeln('  Operations:');
-
-            foreach ($stack->getStackOperations() as $operation) {
-                $output->write('    '.$operation->name.': ');
-                $output->writeln($this->rokkaHelper->formatStackOperationOptions($operation->options));
-            }
-        }
-
-        $options = $stack->getStackOptions();
-        if (!empty($options)) {
-            $output->writeln('  Options:');
-            foreach ($stack->getStackOptions() as $name => $value) {
-                $output->write('    '.$name.': ');
-                $output->writeln('<info>'.$value.'</info>');
-            }
-        }
-    }
-
-    /**
-     * @param DynamicMetadataInterface $metadata
-     *
-     * @return string
-     */
-    protected static function getDynamicMetadataInfo(DynamicMetadataInterface $metadata)
-    {
-        $info = null;
-        switch ($metadata->getName()) {
-            case 'SubjectArea':
-                $data = [];
-                /* @var SubjectArea $metadata */
-                foreach (['x', 'y', 'width', 'height'] as $property) {
-                    $data[] = $property.':'.$metadata->$property;
-                }
-                $info = implode('|', $data);
-                break;
-        }
-
-        return $info;
     }
 }
