+++ src/Command/OrganizationMembershipAddCommand.php
@@ -37,7 +37,7 @@ class OrganizationMembershipAddCommand extends BaseRokkaCliCommand
             $membership = $client->getMembership($organizationName, $email);
 
             $output->writeln('Membership');
-            self::outputOrganizationMembershipInfo($membership, $output);
+            $this->formatterHelper->outputOrganizationMembershipInfo($membership, $output);
         }
 
         return 0;
