+++ src/Command/StackInfoCommand.php
@@ -13,8 +13,8 @@ class StackInfoCommand extends BaseRokkaCliCommand
     {
         $this
             ->setName('stack:info')
-            ->setDescription('Delete the given Stack from an Organization.')
-            ->addArgument('stack-name', InputArgument::REQUIRED, 'The Stack name to load')
+            ->setDescription('Print information about the specified Stack')
+            ->addArgument('stack-name', InputArgument::REQUIRED, 'The Stack name to show')
             ->addOption('organization', null, InputOption::VALUE_REQUIRED, 'The organization to load the Stacks from')
         ;
     }
@@ -35,7 +35,7 @@ class StackInfoCommand extends BaseRokkaCliCommand
 
         $stack = $client->getStack($stackName, $organization);
 
-        $this->outputStackInfo($stack, $output);
+        $this->formatterHelper->outputStackInfo($stack, $output);
 
         return 0;
     }
